# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Competitor Insight Hub API",
    "description": "API для мониторинга новостей из мира ИИ-индустрии",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/competitors/compare": {
      "post": {
        "tags": [
          "Competitors"
        ],
        "summary": "Compare Companies",
        "description": "Compare multiple companies\n\nRequest body:\n{\n    \"company_ids\": [\"uuid1\", \"uuid2\", \"uuid3\"],\n    \"date_from\": \"2025-01-01\",  // optional\n    \"date_to\": \"2025-01-31\",     // optional\n    \"name\": \"Q1 2025 Comparison\" // optional\n}",
        "operationId": "compare_companies_api_v1_competitors_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/competitors/comparisons": {
      "get": {
        "tags": [
          "Competitors"
        ],
        "summary": "Get User Comparisons",
        "description": "Get user's saved comparisons",
        "operationId": "get_user_comparisons_api_v1_competitors_comparisons_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/competitors/comparisons/{comparison_id}": {
      "get": {
        "tags": [
          "Competitors"
        ],
        "summary": "Get Comparison",
        "description": "Get specific comparison details",
        "operationId": "get_comparison_api_v1_competitors_comparisons__comparison_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Competitors"
        ],
        "summary": "Delete Comparison",
        "description": "Delete a comparison",
        "operationId": "delete_comparison_api_v1_competitors_comparisons__comparison_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/competitors/activity/{company_id}": {
      "get": {
        "tags": [
          "Competitors"
        ],
        "summary": "Get Company Activity",
        "description": "Get activity metrics for a specific company",
        "operationId": "get_company_activity_api_v1_competitors_activity__company_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/competitors/suggest/{company_id}": {
      "get": {
        "tags": [
          "Competitors"
        ],
        "summary": "Suggest Competitors",
        "description": "Автоматически подобрать конкурентов\n\nQuery params:\n    limit: сколько конкурентов вернуть (default: 5, max: 10)\n    days: за какой период анализировать (default: 30)",
        "operationId": "suggest_competitors_api_v1_competitors_suggest__company_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/competitors/themes": {
      "post": {
        "tags": [
          "Competitors"
        ],
        "summary": "Analyze Themes",
        "description": "Анализ новостных тем для списка компаний\n\nBody:\n    {\n        \"company_ids\": [\"uuid1\", \"uuid2\", \"uuid3\"],\n        \"date_from\": \"2025-01-01\",  // optional\n        \"date_to\": \"2025-01-31\"     // optional\n    }",
        "operationId": "analyze_themes_api_v1_competitors_themes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/competitors/changes/{company_id}": {
      "get": {
        "tags": [
          "Competitors"
        ],
        "summary": "List Change Events",
        "operationId": "list_change_events_api_v1_competitors_changes__company_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/competitors/changes/{event_id}/recompute": {
      "post": {
        "tags": [
          "Competitors"
        ],
        "summary": "Recompute Change Event",
        "operationId": "recompute_change_event_api_v1_competitors_changes__event_id__recompute_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Competitor not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}